use cast_vector<uint64_t>

+++ inference-engine/src/mkldnn_plugin/mkldnn_plugin.cpp
@@ -279,10 +279,12 @@ static void Transformation(CNNNetwork& clonedNetwork, const Config& conf) {
             });
 
     auto normalizeL2FusionCallback = [](const_node_ptr &node) -> bool {
+        // This is a restriction based on MKLDNNNormalizeL2Node::isSupportedOperation in CPU plugin.
+        // Since CPU can't handle axes other that [1] or [1, ..] we need to keep NormalizeL2 decomposed
         auto axes_node = std::dynamic_pointer_cast<const ngraph::opset4::Constant>(node);
         if (!axes_node)
             return true;
-        auto axes = axes_node->cast_vector<size_t>();
+        auto axes = axes_node->cast_vector<uint64_t>();
         if (axes.size() == 1 && axes[0] == 1) {
             return false;
         }
